daemon: extract map sweeper lifecycle into hive cell
This commit extracts the lifecycle aspect of the map sweeper
(cleanup stale and disabled maps after endpoint restoration) into
a new Hive Cell in the package `pkg/datapath/maps`).
