+++ daemon/cmd/datapath.go
@@ -6,13 +6,11 @@ package cmd
 import (
 	"fmt"
 	"log/slog"
-	"os"
 	"strings"
 
 	"github.com/vishvananda/netlink"
 
 	"github.com/cilium/cilium/pkg/datapath/linux/safenetlink"
-	"github.com/cilium/cilium/pkg/endpointmanager"
 	"github.com/cilium/cilium/pkg/logging/logfields"
 	"github.com/cilium/cilium/pkg/maps/ctmap"
 	"github.com/cilium/cilium/pkg/maps/fragmap"
@@ -21,7 +19,6 @@ import (
 	"github.com/cilium/cilium/pkg/maps/metricsmap"
 	"github.com/cilium/cilium/pkg/maps/nat"
 	"github.com/cilium/cilium/pkg/maps/neighborsmap"
-	"github.com/cilium/cilium/pkg/maps/policymap"
 	"github.com/cilium/cilium/pkg/option"
 )
 
@@ -83,56 +80,6 @@ func clearCiliumVeths(logger *slog.Logger) error {
 		}
 	}
 	return nil
-}
-
-// EndpointMapManager is a wrapper around an endpointmanager as well as the
-// filesystem for removing maps related to endpoints from the filesystem.
-type EndpointMapManager struct {
-	logger *slog.Logger
-	endpointmanager.EndpointManager
-}
-
-// RemoveDatapathMapping unlinks the endpointID from the global policy map, preventing
-// packets that arrive on this node from being forwarded to the endpoint that
-// used to exist with the specified ID.
-func (e *EndpointMapManager) RemoveDatapathMapping(endpointID uint16) error {
-	return policymap.RemoveGlobalMapping(e.logger, uint32(endpointID))
-}
-
-// RemoveMapPath removes the specified path from the filesystem.
-func (e *EndpointMapManager) RemoveMapPath(path string) {
-	if err := os.RemoveAll(path); err != nil {
-		e.logger.Warn(
-			"Error while deleting stale map file",
-			logfields.Path, path,
-		)
-	} else {
-		e.logger.Info(
-			"Removed stale bpf map",
-			logfields.Path, path,
-		)
-	}
-}
-
-// ListMapsDir gives names of files (or subdirectories) found in the specified path.
-func (e *EndpointMapManager) ListMapsDir(path string) []string {
-	var maps []string
-
-	entries, err := os.ReadDir(path)
-	if err != nil {
-		e.logger.Warn(
-			"Error while listing maps dir",
-			logfields.Path, path,
-			logfields.Error, err,
-		)
-		return maps
-	}
-
-	for _, e := range entries {
-		maps = append(maps, e.Name())
-	}
-
-	return maps
 }
 
 // initMaps opens all BPF maps (and creates them if they do not exist). This
